refactor: separate chain URL extraction
