+++ tests/test_cache.py
@@ -1,7 +1,8 @@
 from unittest.mock import patch
 
-from blockscout_mcp_server.cache import ChainCache, find_blockscout_url
+from blockscout_mcp_server.cache import ChainCache
 from blockscout_mcp_server.config import config
+from blockscout_mcp_server.tools.common import find_blockscout_url
 
 
 def test_find_blockscout_url_success():
@@ -22,7 +23,7 @@ def test_find_blockscout_url_no_match():
 def test_chain_cache_basic_flow():
     cache = ChainCache()
     with patch("time.time", return_value=1000):
-        cache.set("1", {"explorers": [{"hostedBy": "blockscout", "url": "https://a"}]})
+        cache.set("1", "https://a")
     assert cache.get("1") == ("https://a", 1000 + config.chain_cache_ttl_seconds)
 
 
@@ -35,12 +36,12 @@ def test_chain_cache_set_failure():
 
 def test_chain_cache_bulk_set():
     cache = ChainCache()
-    chains = {
-        "1": {"explorers": [{"hostedBy": "blockscout", "url": "https://a"}]},
-        "2": {"explorers": [{"hostedBy": "blockscout", "url": "https://b"}]},
+    chain_urls = {
+        "1": "https://a",
+        "2": "https://b",
     }
     with patch("time.time", return_value=3000):
-        cache.bulk_set(chains)
+        cache.bulk_set(chain_urls)
     assert cache.get("1") == ("https://a", 3000 + config.chain_cache_ttl_seconds)
     assert cache.get("2") == ("https://b", 3000 + config.chain_cache_ttl_seconds)
 
@@ -48,7 +49,7 @@ def test_chain_cache_bulk_set():
 def test_chain_cache_invalidate():
     cache = ChainCache()
     with patch("time.time", return_value=4000):
-        cache.set("1", {"explorers": [{"hostedBy": "blockscout", "url": "https://a"}]})
+        cache.set("1", "https://a")
     assert cache.get("1") == ("https://a", 4000 + config.chain_cache_ttl_seconds)
     cache.invalidate("1")
     assert cache.get("1") is None
